+++ src/routes/cliente.index.tsx
@@ -1,5 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 import { ClientVitrine } from "@/components/cliente/ClientVitrine";
+import { InstallAppButton } from "@/components/cliente/InstallAppButton";
 import { SkeletonScreen } from "@/components/WelcomeSplash";
 import { useCustomerSales } from "@/lib/customer-queries";
 
@@ -16,8 +17,13 @@ export function Page_cliente_index() {
 
   return (
     <div className="flex flex-col h-full min-h-0">
-      <div className="flex-1 min-h-0 px-4 pb-4">
-        <ClientVitrine />
+      <div className="flex-1 min-h-0 px-4 pb-4 flex flex-col gap-3">
+        <div className="flex-1 min-h-0">
+          <ClientVitrine />
+        </div>
+        <div className="shrink-0">
+          <InstallAppButton />
+        </div>
       </div>
     </div>
   );